mv file and add softlink
